MVP: use manifold-connection for auth
