+++ src/components/manifold-performance/manifold-performance.tsx
@@ -13,7 +13,7 @@ const loggableEvents = [
   'manifold-graphql-fetch-duration',
   'manifold-error',
   'manifold-time-to-render',
-  'receiveManifoldToken',
+  'manifold-token-receive',
 ];
 
 @Component({ tag: 'manifold-performance' })
@@ -38,20 +38,15 @@ export class ManifoldPerformance {
   };
 
   logMetric = (e: CustomEvent) => {
-    if (e.type === 'receiveManifoldToken' && !e.detail.token) {
-      // Only log duration if token is defined
-      return;
-    }
-    if (e.type === 'receiveManifoldToken') {
+    if (e.type === 'manifold-token-receive') {
       delete e.detail.token;
-      delete e.detail.expiry;
     }
     if (!this.ddLogs) {
       this.logQueue.push(e);
       return;
     }
 
-    if (e.type === 'manifold-error' || (e.type === 'receiveManifoldToken' && e.detail.error)) {
+    if (e.type === 'manifold-error') {
       this.ddLogs.logger.error(e.type, { type: e.type, ...e.detail });
     } else {
       this.ddLogs.logger.info(e.type, { type: e.type, ...e.detail });
