+++ src/types/auth.ts
@@ -3,13 +3,6 @@ export interface AuthError {
   message: string;
 }
 
-export interface AuthToken {
-  token?: string;
-  expiry?: number;
-  error?: AuthError;
-  duration?: number;
-}
-
 export interface PumaAuthToken {
   access_token?: string;
   expiry?: number;
